feat(ledger): add ledger-common module with Transaction, TransactionValidator, ExceptionMessages; update poms, imports and entity/component scans
