+++ src/ledger/balancereader/src/main/java/anthos/samples/bankofanthos/balancereader/LedgerReader.java
@@ -16,6 +16,7 @@
 
 package anthos.samples.bankofanthos.balancereader;
 
+import anthos.samples.bankofanthos.ledger.common.Transaction;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
