+++ src/ledger/balancereader/src/main/java/anthos/samples/bankofanthos/balancereader/TransactionRepository.java
@@ -14,6 +14,7 @@
 
 package anthos.samples.bankofanthos.balancereader;
 
+import anthos.samples.bankofanthos.ledger.common.Transaction;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.Query;
