+++ src/ledger/ledgerwriter/src/main/java/anthos/samples/bankofanthos/ledgerwriter/LedgerWriterApplication.java
@@ -28,7 +28,9 @@ import org.apache.logging.log4j.Logger;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.actuate.autoconfigure.tracing.zipkin.ZipkinAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.web.client.RestTemplate;
 
 /**
@@ -37,6 +39,13 @@ import org.springframework.web.client.RestTemplate;
  * Microservice to accept new transactions for the bank ledger.
  */
 @SpringBootApplication(exclude = ZipkinAutoConfiguration.class)
+@ComponentScan(basePackages = {
+    "anthos.samples.bankofanthos.ledgerwriter",
+    "anthos.samples.bankofanthos.ledger.common"
+})
+@EntityScan(basePackages = {
+    "anthos.samples.bankofanthos.ledger.common"
+})
 public class LedgerWriterApplication {
 
     private static final Logger LOGGER =
