+++ src/ledger/transactionhistory/src/main/java/anthos/samples/bankofanthos/transactionhistory/LedgerReader.java
@@ -16,6 +16,8 @@
 
 package anthos.samples.bankofanthos.transactionhistory;
 
+import anthos.samples.bankofanthos.ledger.common.Transaction;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
